More implementations for image inference

+++ src/main.cpp
@@ -2,15 +2,15 @@
 
 extern "C" {
 
-unsigned long constexpr IMAGE_RESILUTION = 448;
-unsigned long constexpr NUM_CHANNELS = 3;
-unsigned long constexpr NUM_CLASSES = 3;
+unsigned int constexpr IMAGE_RESILUTION = 448;
+unsigned int constexpr NUM_CHANNELS = 3;
+unsigned int constexpr NUM_CLASSES = 3;
 
-unsigned long constexpr SIZE_Y = IMAGE_RESILUTION;
-unsigned long constexpr SIZE_X = IMAGE_RESILUTION;
-unsigned long constexpr SIZE_C = NUM_CHANNELS;
+unsigned int constexpr SIZE_Y = IMAGE_RESILUTION;
+unsigned int constexpr SIZE_X = IMAGE_RESILUTION;
+unsigned int constexpr SIZE_C = NUM_CHANNELS;
 
-unsigned long constexpr SIZE_O = NUM_CLASSES;
+unsigned int constexpr SIZE_O = NUM_CLASSES;
 
 typedef float outtype;
